+++ cli/azd/pkg/apphost/generate_test.go
@@ -3,10 +3,12 @@ package apphost
 import (
 	"context"
 	_ "embed"
+	"fmt"
 	"io/fs"
 	"os"
 	"path/filepath"
 	"runtime"
+	"strings"
 	"testing"
 
 	"github.com/azure/azure-dev/cli/azd/pkg/exec"
@@ -23,6 +25,9 @@ var aspireDockerManifest []byte
 
 //go:embed testdata/aspire-storage.json
 var aspireStorageManifest []byte
+
+//go:embed testdata/aspire-args.json
+var aspireArgsManifest []byte
 
 //go:embed testdata/aspire-bicep.json
 var aspireBicepManifest []byte
@@ -72,10 +77,9 @@ func TestAspireEscaping(t *testing.T) {
 
 	m, err := ManifestFromAppHost(ctx, filepath.Join("testdata", "AspireDocker.AppHost.csproj"), mockCli, "")
 	require.NoError(t, err)
-
 	for _, name := range []string{"api"} {
 		t.Run(name, func(t *testing.T) {
-			tmpl, err := ContainerAppManifestTemplateForProject(m, name)
+			tmpl, err := ContainerAppManifestTemplateForProject(m, name, AppHostOptions{})
 			require.NoError(t, err)
 			snapshot.SnapshotT(t, tmpl)
 		})
@@ -95,7 +99,7 @@ func TestAspireStorageGeneration(t *testing.T) {
 	m, err := ManifestFromAppHost(ctx, filepath.Join("testdata", "AspireDocker.AppHost.csproj"), mockCli, "")
 	require.NoError(t, err)
 
-	files, err := BicepTemplate(m)
+	files, err := BicepTemplate(m, AppHostOptions{})
 	require.NoError(t, err)
 
 	err = fs.WalkDir(files, ".", func(path string, d fs.DirEntry, err error) error {
@@ -137,7 +141,7 @@ func TestAspireBicepGeneration(t *testing.T) {
 	m, err := ManifestFromAppHost(ctx, filepath.Join("testdata", "AspireDocker.AppHost.csproj"), mockCli, "")
 	require.NoError(t, err)
 
-	files, err := BicepTemplate(m)
+	files, err := BicepTemplate(m, AppHostOptions{})
 	require.NoError(t, err)
 
 	err = fs.WalkDir(files, ".", func(path string, d fs.DirEntry, err error) error {
@@ -160,7 +164,7 @@ func TestAspireBicepGeneration(t *testing.T) {
 
 	for _, name := range []string{"frontend"} {
 		t.Run(name, func(t *testing.T) {
-			tmpl, err := ContainerAppManifestTemplateForProject(m, name)
+			tmpl, err := ContainerAppManifestTemplateForProject(m, name, AppHostOptions{})
 			require.NoError(t, err)
 			snapshot.SnapshotT(t, tmpl)
 		})
@@ -180,23 +184,15 @@ func TestAspireDockerGeneration(t *testing.T) {
 	m, err := ManifestFromAppHost(ctx, filepath.Join("testdata", "AspireDocker.AppHost.csproj"), mockCli, "")
 	require.NoError(t, err)
 
-	// The App Host manifest does not set the external bit for project resources. Instead, `azd` or whatever tool consumes
-	// the manifest should prompt the user to select which services should be exposed. For this test, we manually set the
-	// external bit on the resources on the webfrontend resource to simulate the user selecting the webfrontend to be
-	// exposed.
-	for _, value := range m.Resources["nodeapp"].Bindings {
-		value.External = true
-	}
-
-	for _, name := range []string{"nodeapp"} {
+	for _, name := range []string{"nodeapp", "api"} {
 		t.Run(name, func(t *testing.T) {
-			tmpl, err := ContainerAppManifestTemplateForProject(m, name)
+			tmpl, err := ContainerAppManifestTemplateForProject(m, name, AppHostOptions{})
 			require.NoError(t, err)
 			snapshot.SnapshotT(t, tmpl)
 		})
 	}
 
-	files, err := BicepTemplate(m)
+	files, err := BicepTemplate(m, AppHostOptions{})
 	require.NoError(t, err)
 
 	err = fs.WalkDir(files, ".", func(path string, d fs.DirEntry, err error) error {
@@ -218,6 +214,60 @@ func TestAspireDockerGeneration(t *testing.T) {
 	require.NoError(t, err)
 }
 
+func TestAspireDashboardGeneration(t *testing.T) {
+	if runtime.GOOS == "windows" {
+		t.Skip("Skipping due to EOL issues on Windows with the baselines")
+	}
+
+	ctx := context.Background()
+	mockCtx := mocks.NewMockContext(ctx)
+	mockPublishManifest(mockCtx, aspireDockerManifest, nil)
+	mockCli := dotnet.NewDotNetCli(mockCtx.CommandRunner)
+
+	m, err := ManifestFromAppHost(ctx, filepath.Join("testdata", "AspireDocker.AppHost.csproj"), mockCli, "")
+	require.NoError(t, err)
+
+	files, err := BicepTemplate(m, AppHostOptions{AspireDashboard: true})
+	require.NoError(t, err)
+
+	err = fs.WalkDir(files, ".", func(path string, d fs.DirEntry, err error) error {
+		if err != nil {
+			return err
+		}
+		if d.IsDir() {
+			return nil
+		}
+		contents, err := fs.ReadFile(files, path)
+		if err != nil {
+			return err
+		}
+		t.Run(path, func(t *testing.T) {
+			snapshot.SnapshotT(t, string(contents))
+		})
+		return nil
+	})
+	require.NoError(t, err)
+}
+
+func TestAspireArgsGeneration(t *testing.T) {
+	if runtime.GOOS == "windows" {
+		t.Skip("Skipping due to EOL issues on Windows with the baselines")
+	}
+
+	ctx := context.Background()
+	mockCtx := mocks.NewMockContext(ctx)
+	mockPublishManifest(mockCtx, aspireArgsManifest, nil)
+	mockCli := dotnet.NewDotNetCli(mockCtx.CommandRunner)
+
+	m, err := ManifestFromAppHost(ctx, filepath.Join("testdata", "AspireArgs.AppHost.csproj"), mockCli, "")
+	require.NoError(t, err)
+
+	manifest, err := ContainerAppManifestTemplateForProject(m, "apiservice", AppHostOptions{})
+	require.NoError(t, err)
+
+	snapshot.SnapshotT(t, manifest)
+}
+
 func TestAspireContainerGeneration(t *testing.T) {
 	if runtime.GOOS == "windows" {
 		t.Skip("Skipping due to EOL issues on Windows with the baselines")
@@ -231,7 +281,7 @@ func TestAspireContainerGeneration(t *testing.T) {
 	m, err := ManifestFromAppHost(ctx, filepath.Join("testdata", "AspireDocker.AppHost.csproj"), mockCli, "")
 	require.NoError(t, err)
 
-	files, err := BicepTemplate(m)
+	files, err := BicepTemplate(m, AppHostOptions{})
 	require.NoError(t, err)
 
 	err = fs.WalkDir(files, ".", func(path string, d fs.DirEntry, err error) error {
@@ -356,14 +406,17 @@ func TestInjectValueForBicepParameter(t *testing.T) {
 	require.Equal(t, expectedParameter, value)
 	require.False(t, inject)
 
-	expectedParameter = "resources.outputs.SERVICE_BINDING_EXAMPLEKV_NAME"
+	uniqueName := strings.ToUpper("kv" + uniqueFnvNumber(resourceName))
+	expectedParameter = fmt.Sprintf("resources.outputs.SERVICE_BINDING_%s_NAME", uniqueName)
 	value, inject, err = injectValueForBicepParameter(resourceName, param, "")
 	require.NoError(t, err)
 	require.Equal(t, expectedParameter, value)
 	require.True(t, inject)
 
-	expectedParameter = "resources.outputs.SERVICE_BINDING_EXAMPLE_01KV_NAME"
-	value, inject, err = injectValueForBicepParameter(resourceName+"-01", param, "")
+	withDash := resourceName + "-01"
+	uniqueName = strings.ToUpper("kv" + uniqueFnvNumber(withDash))
+	expectedParameter = fmt.Sprintf("resources.outputs.SERVICE_BINDING_%s_NAME", uniqueName)
+	value, inject, err = injectValueForBicepParameter(withDash, param, "")
 	require.NoError(t, err)
 	require.Equal(t, expectedParameter, value)
 	require.True(t, inject)