+++ cli/azd/pkg/apphost/generate_types.go
@@ -1,5 +1,7 @@
 package apphost
 
+import "github.com/azure/azure-dev/cli/azd/pkg/custommaps"
+
 type genAppInsight struct{}
 
 type genStorageAccount struct {
@@ -37,17 +39,27 @@ type genContainerApp struct {
 	Volumes []*Volume
 }
 
+type genContainerAppIngressPort struct {
+	External   bool
+	TargetPort int
+}
+
+type genContainerAppIngressAdditionalPortMappings struct {
+	genContainerAppIngressPort
+	ExposedPort int
+}
+
 type genContainerAppIngress struct {
-	External      bool
-	TargetPort    int
-	Transport     string
-	AllowInsecure bool
+	genContainerAppIngressPort
+	Transport              string
+	AllowInsecure          bool
+	AdditionalPortMappings []genContainerAppIngressAdditionalPortMappings
 }
 
 type genContainer struct {
 	Image    string
 	Env      map[string]string
-	Bindings map[string]*Binding
+	Bindings custommaps.WithOrder[Binding]
 	Inputs   map[string]Input
 	Volumes  []*Volume
 }
@@ -56,14 +68,16 @@ type genDockerfile struct {
 	Path      string
 	Context   string
 	Env       map[string]string
-	Bindings  map[string]*Binding
+	Bindings  custommaps.WithOrder[Binding]
 	BuildArgs map[string]string
+	Args      []string
 }
 
 type genProject struct {
 	Path     string
 	Env      map[string]string
-	Bindings map[string]*Binding
+	Args     []string
+	Bindings custommaps.WithOrder[Binding]
 }
 
 type genAppConfig struct{}
@@ -116,6 +130,7 @@ type genBicepTemplateContext struct {
 	HasLogAnalyticsWorkspace        bool
 	RequiresPrincipalId             bool
 	RequiresStorageVolume           bool
+	AspireDashboard                 bool
 	AppInsights                     map[string]genAppInsight
 	ServiceBuses                    map[string]genServiceBus
 	StorageAccounts                 map[string]genStorageAccount
@@ -135,12 +150,14 @@ type genBicepTemplateContext struct {
 }
 
 type genContainerAppManifestTemplateContext struct {
-	Name            string
-	Ingress         *genContainerAppIngress
-	Env             map[string]string
-	Secrets         map[string]string
-	KeyVaultSecrets map[string]string
-	Dapr            *genContainerAppManifestTemplateContextDapr
+	Name                        string
+	Ingress                     *genContainerAppIngress
+	Env                         map[string]string
+	Secrets                     map[string]string
+	KeyVaultSecrets             map[string]string
+	Dapr                        *genContainerAppManifestTemplateContextDapr
+	AutoConfigureDataProtection bool
+	Args                        []string
 }
 
 type genProjectFileContext struct {